Simplify TestPaths class.

+++ src/chezmoi_mousse/_test_paths.py
@@ -1,5 +1,5 @@
 import random
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from enum import StrEnum
 from pathlib import Path
 
@@ -20,135 +20,161 @@ class ProblemChars(StrEnum):
     ZWS = "\u200b"
 
 
-@dataclass(frozen=True, slots=True)
-class Files:
+@dataclass(frozen=True, slots=True, kw_only=True)
+class PileNames:
     BINARY = "_test_file_binary.bin"
     LARGE = "_test_file_large.txt"
-    TOML_APPLY_A = "_test_file_apply_a.toml"
-    TOML_APPLY_D = "_test_file_apply_d.toml"
-    TOML_APPLY_M = "_test_file_apply_m.toml"
-    TOML_APPLY_S = "_test_file_apply_space.toml"
-    TOML_RE_ADD_A = "_test_file_re_add_a.toml"
-    TOML_RE_ADD_D = "_test_file_re_add_d.toml"
-    TOML_RE_ADD_M = "_test_file_re_add_m.toml"
-    TOML_RE_ADD_S = "_test_file_re_add_space.toml"
+    TEST_FILE_1 = "_test_file_1.toml"
+    TEST_FILE_2 = "_test_file_2.toml"
+    TEST_FILE_3 = "_test_file_3.toml"
+    TEST_FILE_4 = "_test_file_4.toml"
     TRICKY_UTF8 = "_test_file_tricky_utf8.txt"
-    UNCHANGED_1 = "_test_file_unchanged_1.toml"
-    UNCHANGED_2 = "_test_file_unchanged_2.toml"
-    UNCHANGED_3 = "_test_file_unchanged_3.toml"
+
+    @property
+    def all_toml_file_names(self) -> list[str]:
+        return [self.TEST_FILE_1, self.TEST_FILE_2, self.TEST_FILE_3, self.TEST_FILE_4]
+
+    @property
+    def toml_files_for_diffs(self) -> list[str]:
+        return [self.TEST_FILE_1, self.TEST_FILE_3]
 
 
 @dataclass(frozen=True, slots=True, kw_only=True)
-class DirNames:
-    HOME = "_test_home"
+class AllTestPaths:
+    DEST_DIR: Path
+    # dir names
+    DIR_WITH_STATUS = "_test_dir_with_status"
     EMPTY = "_test_empty_dir"
-    WITH_STATUS = "_test_dir_with_status"
-    WITH_STATUS_CHILDREN = "_test_dir_with_status_children"
     SUB_DIR = "_test_sub_dir"
-    TEST_DIR = "_test_dir"
-    UNCHANGED = "_test_dir_unchanged"
-    CHANGED = "_test_dir_changed"
+    WITH_NESTED_STATUS_CHILDREN = "_test_with_status_children"
+    WITH_SPECIAL_FILES = "_test_special_files"
+    WITHOUT_STATUS_CHILDREN = "_test_with_status_children"
+
+    # file names
+    file_names = PileNames()
+
+    @property
+    def dir_with_special_files(self) -> Path:
+        return self.DEST_DIR / self.WITH_SPECIAL_FILES
+
+    @property
+    def dir_with_nested_dir_with_status_children(self) -> Path:
+        return self.DEST_DIR / self.WITH_NESTED_STATUS_CHILDREN
+
+    @property
+    def dir_with_status(self) -> Path:
+        return self.DEST_DIR / self.DIR_WITH_STATUS
+
+    @property
+    def dir_without_status_children(self) -> Path:
+        return self.dir_with_nested_dir_with_status_children / self.SUB_DIR
+
+    @property
+    def empty_dir_path(self) -> Path:
+        return self.DEST_DIR / self.EMPTY
+
+    @property
+    def nested_dir_with_status_paths(self) -> Path:
+        return self.DEST_DIR / self.WITH_NESTED_STATUS_CHILDREN / self.SUB_DIR
+
+    @property
+    def sub_dir_path(self) -> Path:
+        return self.DEST_DIR / self.SUB_DIR
+
+    @property
+    def all_dirs_to_create(self) -> list[Path]:
+        # to create the dirs with parents True
+        return [
+            self.dir_with_special_files,
+            self.dir_with_status,
+            self.dir_without_status_children,
+            self.empty_dir_path,
+            self.nested_dir_with_status_paths,
+            self.sub_dir_path,
+        ]
+
+    @property
+    def dirs_with_toml_files(self) -> list[Path]:
+        return [
+            self.DEST_DIR,
+            self.dir_without_status_children,
+            self.nested_dir_with_status_paths,
+            self.sub_dir_path,
+            self.dir_with_status,
+        ]
+
+    @property
+    def toml_files_to_create(self) -> list[Path]:
+        to_create: list[Path] = []
+        for file_name in self.file_names.all_toml_file_names:
+            for dir in self.dirs_with_toml_files:
+                to_create.append(dir / file_name)
+        return to_create
+
+    @property
+    def toml_files_for_diff(self) -> list[Path]:
+        to_diff: list[Path] = []
+        dirs_to_exclude = [
+            self.dir_with_nested_dir_with_status_children,
+            self.dir_with_special_files,
+            self.dir_without_status_children,
+            self.empty_dir_path,
+            self.nested_dir_with_status_paths,
+        ]
+        for file_name in self.file_names.toml_files_for_diffs:
+            for dir in self.dirs_with_toml_files:
+                if dir in dirs_to_exclude:
+                    continue
+                to_diff.append(dir / file_name)
+        return to_diff
+
+    @property
+    def large_file_path(self) -> Path:
+        return self.DEST_DIR / self.WITH_SPECIAL_FILES / self.file_names.LARGE
+
+    @property
+    def binary_file_path(self) -> Path:
+        return self.DEST_DIR / self.WITH_SPECIAL_FILES / self.file_names.BINARY
+
+    @property
+    def tricky_utf8_file_path(self) -> Path:
+        return self.DEST_DIR / self.WITH_SPECIAL_FILES / self.file_names.TRICKY_UTF8
+
+    @property
+    def all_test_paths(self) -> list[Path]:
+        return sorted(
+            self.all_dirs_to_create
+            + self.toml_files_to_create
+            + [self.large_file_path, self.binary_file_path, self.tricky_utf8_file_path]
+        )
 
 
 @dataclass(slots=True, kw_only=True)
 class TestPaths:
-    files: Files = Files()
-    dir_names: DirNames = DirNames()
-    home_dir: Path = Path.home() / dir_names.HOME
-    test_dir: Path = home_dir / dir_names.TEST_DIR
-    status_children_dir: Path = (
-        test_dir / dir_names.WITH_STATUS_CHILDREN / dir_names.SUB_DIR
-    )
-    binary_file_path: Path = test_dir / files.BINARY
-    large_file_path: Path = test_dir / files.LARGE
-    tricky_utf8_file_path: Path = test_dir / files.TRICKY_UTF8
-    dir_paths: set[Path] = field(default_factory=lambda: set())
-    file_paths: set[Path] = field(default_factory=lambda: set())
+    all_paths: AllTestPaths = AllTestPaths(DEST_DIR=Path.home())
 
-    @property
-    def _changed_file_names(self) -> list[str]:
-        return [
-            self.files.TOML_APPLY_A,
-            self.files.TOML_APPLY_D,
-            self.files.TOML_APPLY_M,
-            self.files.TOML_APPLY_S,
-            self.files.TOML_RE_ADD_A,
-            self.files.TOML_RE_ADD_D,
-            self.files.TOML_RE_ADD_M,
-            self.files.TOML_RE_ADD_S,
-        ]
+    def create_dir_paths(self) -> list[str]:
+        created_dirs: list[str] = []
+        for dir in self.all_paths.all_dirs_to_create:
+            if not dir.exists():
+                dir.mkdir(parents=True, exist_ok=True)
+                created_dirs.append(str(dir))
+        return created_dirs
 
-    @property
-    def _unchanged_file_names(self) -> list[str]:
-        return [self.files.UNCHANGED_1, self.files.UNCHANGED_2, self.files.UNCHANGED_3]
-
-    @property
-    def _test_dir_paths(self) -> list[Path]:
-        return [
-            self.home_dir,
-            self.home_dir / self.dir_names.UNCHANGED,
-            self.test_dir,
-            self.test_dir / self.dir_names.UNCHANGED,
-            self.test_dir / self.dir_names.CHANGED,
-            self.test_dir / self.dir_names.EMPTY,
-            self.test_dir / self.dir_names.WITH_STATUS_CHILDREN,
-            self.test_dir
-            / self.dir_names.WITH_STATUS_CHILDREN
-            / self.dir_names.SUB_DIR,
-        ]
-
-    @property
-    def _all_existing_paths(self) -> list[Path]:
-        paths = sorted(self.dir_paths | self.file_paths)
-        return [p for p in paths if p.exists()]
-
-    def create_paths_on_disk(self) -> list[str]:
-        created_paths: list[str] = []
-        for dir in self.dir_paths:
-            dir.mkdir(parents=True, exist_ok=True)
-            created_paths.append(str(dir))
-        created_paths.append(self._create_tricky_utf8_file())
-        created_paths.append(self._create_large_file())
-        created_paths.append(self._create_binary_file())
-        for file in self.file_paths:
-            if file.exists() or file.parent.name == self.dir_names.EMPTY:
-                continue
-            if file.name.endswith(".toml"):
-                doc = self._get_fake_toml_document()
-                with open(file, "w", encoding="utf-8") as f:
-                    f.write(doc.as_string())
-                created_paths.append(str(file))
-        return created_paths
-
-    def _get_fake_toml_document(self) -> tomlkit.TOMLDocument:
-        doc = tomlkit.document()
-        doc["title"] = FAKER.sentence(nb_words=6)
-        doc["version"] = FAKER.pyfloat(left_digits=1, right_digits=2, positive=True)
-        doc["debug"] = FAKER.boolean()
-        doc["hosts"] = [FAKER.hostname() for _ in range(10)]
-        doc["ports"] = [FAKER.port_number() for _ in range(10)]
-        some_table = tomlkit.table()
-        some_table["id"] = FAKER.uuid4()
-        some_table["date"] = FAKER.date_time().isoformat()
-        some_table["text"] = FAKER.paragraph(nb_sentences=12)
-        doc["some_table"] = some_table
-        return doc
-
-    def _create_binary_file(self) -> str:
-        file = self.binary_file_path
+    def create_binary_file(self) -> str:
+        file = self.all_paths.binary_file_path
         content = FAKER.binary(length=2048)
         with open(file, "wb") as f:
             f.write(content)
         return str(file)
 
-    def _create_large_file(self) -> str:
-        file = self.large_file_path
+    def create_large_file(self) -> str:
         content = FAKER.text(max_nb_chars=700000)
-        with open(file, "w", encoding="utf-8") as f:
+        with open(self.all_paths.large_file_path, "w", encoding="utf-8") as f:
             f.write(content)
-        return str(file)
+        return str(self.all_paths.large_file_path)
 
-    def _create_tricky_utf8_file(self) -> str:
+    def create_tricky_utf8_file(self) -> str:
         parts: list[str] = []
         parts.append(FAKER.sentence(nb_words=6))
         parts.append(
@@ -176,32 +202,68 @@ class TestPaths:
             + FAKER.word()
         )
         parts.append("A" + ProblemChars.ZWJ + "B" + ProblemChars.ZWJ + "C")
-        with open(self.tricky_utf8_file_path, "w", encoding="utf-8") as f:
+        with open(self.all_paths.tricky_utf8_file_path, "w", encoding="utf-8") as f:
             f.write("\n".join(parts))
-        return str(self.tricky_utf8_file_path)
+        return str(self.all_paths.tricky_utf8_file_path)
 
-    def list_existing_test_paths(self) -> str | None:
-        existing_paths = [str(p) for p in self._all_existing_paths if p.exists()]
+    def create_toml_files(self) -> list[str]:
+        def get_fake_toml_string() -> str:
+            doc = tomlkit.document()
+            doc["title"] = FAKER.sentence(nb_words=6)
+            doc["version"] = FAKER.pyfloat(left_digits=1, right_digits=2, positive=True)
+            doc["debug"] = FAKER.boolean()
+            doc["hosts"] = [FAKER.hostname() for _ in range(10)]
+            doc["ports"] = [FAKER.port_number() for _ in range(10)]
+            some_table = tomlkit.table()
+            some_table["id"] = FAKER.uuid4()
+            some_table["date"] = FAKER.date_time().isoformat()
+            some_table["text"] = FAKER.paragraph(nb_sentences=12)
+            doc["some_table"] = some_table
+            return doc.as_string()
+
+        created_files: list[str] = []
+        for file in self.all_paths.toml_files_to_create:
+            if file.exists():
+                file.unlink()
+            with open(file, "w", encoding="utf-8") as f:
+                f.write(get_fake_toml_string())
+            created_files.append(str(file))
+        return created_files
+
+    def create_paths_on_disk(self) -> list[str]:
+        created_paths = self.create_dir_paths()
+        created_paths.append(self.create_binary_file())
+        created_paths.append(self.create_large_file())
+        created_paths.append(self.create_tricky_utf8_file())
+        created_paths.extend(self.create_toml_files())
+        return created_paths
+
+    def list_existing_test_paths(self) -> str | list[Path]:
+        existing_paths = [p for p in self.all_paths.all_test_paths if p.exists()]
         if not existing_paths:
-            return None
-        return "[$text-success]Existing paths:[/]\n" + "\n".join(existing_paths)
+            return "[$text-warning]No test paths exist on disk.[/]"
+        return existing_paths
 
     def remove_test_paths(self) -> str:
         removed_paths: list[str] = []
         # Remove files first
-        for path in self._all_existing_paths:
-            if path.exists() and path.is_file():
+        existing_paths = self.list_existing_test_paths()
+        if isinstance(existing_paths, str):
+            return existing_paths
+        for path in existing_paths:
+            if path.is_file():
                 path.unlink()
                 removed_paths.append(str(path))
         # Then remove directories
-        for dir in sorted(self.dir_paths, reverse=True):
-            dir_path = Path(dir)
+        for dir_path in sorted(self.all_paths.all_dirs_to_create, reverse=True):
             if dir_path.exists():
                 dir_path.rmdir()
                 removed_paths.append(str(dir_path))
         if not removed_paths:
             return "[$text-warning]No test paths to remove.[/]"
-        return "[$text-success]All test paths removed.[/]"
+        removed_paths.sort()
+        removed_path_strings = "\n".join([str(p) for p in removed_paths])
+        return "[$text-success]Removed paths:[/]\n" + removed_path_strings
 
     def create_file_diffs(self) -> str:
         if not self.list_existing_test_paths():
@@ -209,7 +271,7 @@ class TestPaths:
         modified: list[str] = []
 
         # Modify LARGE file
-        large_file_path = self.large_file_path
+        large_file_path = self.all_paths.large_file_path
         if large_file_path.exists():
             with open(large_file_path, "w", encoding="utf-8") as f:
                 modified_content = (
@@ -221,12 +283,8 @@ class TestPaths:
             modified.append(str(large_file_path))
 
         # Modify TOML files
-        for file in self.file_paths:
-            if (
-                file.exists()
-                and file.name not in self._unchanged_file_names
-                and file.suffix == ".toml"
-            ):
+        for file in self.all_paths.toml_files_for_diff:
+            if file.exists():
                 modified_content = (
                     file.read_text(encoding="utf-8")
                     .replace("title", "new_title")
@@ -235,23 +293,9 @@ class TestPaths:
                 with open(file, "w", encoding="utf-8") as f:
                     f.write(modified_content)
                 modified.append(str(file))
-        return "[$text-primary]Modified paths:[/]\n" + "\n".join(modified)
 
-    def __post_init__(self):
-        # populate test paths
-        for dir in self._test_dir_paths:
-            self.dir_paths.add(dir)
-        self.file_paths.add(self.binary_file_path)
-        self.file_paths.add(self.large_file_path)
-        self.file_paths.add(self.tricky_utf8_file_path)
-        for file in self._unchanged_file_names:
-            for dir in self.dir_paths:
-                if (
-                    dir.name in (self.dir_names.HOME, self.dir_names.UNCHANGED)
-                    or dir == self.test_dir / self.dir_names.WITH_STATUS_CHILDREN
-                ):
-                    self.file_paths.add(dir / file)
-        for file in self._changed_file_names:
-            for dir in self.dir_paths:
-                if dir == self.status_children_dir or dir == self.home_dir:
-                    self.file_paths.add(dir / file)
+        modified.append(self.create_binary_file())
+        modified.append(self.create_large_file())
+        modified.append(self.create_tricky_utf8_file())
+
+        return "[$text-primary]Modified paths:[/]\n" + "\n".join(modified)
